Refactor button components to not use inheritance of renderable component
Introduce TextBtn is a simple button with text and description on click (not
implemented yet). It is a smart component. And it renders GenericBtn passing
props to it.

GenericBtn is a dumb component which renders props as is (to be simplified even
more in future)

Previous approach with inheritance caused issues with MobX.